third_party/blink/web_tests/external/wpt/css/css-scoping/has-slotted-functional-changing-001.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Changing content targetting :has-slotted(...) and :has-slotted through a single shadow root
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(0, 0, 255)"
Harness: the test ran to completion.
